+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ## [unreleased]
 
+### Added
+
+- `sandstone/Scroller` prop `aria-label` to be read out instead of a body text
+
 ### Fixed
 
 - `sandstone/Scroller` not to read out thumb audio guidance when focusing on the body